fix some Integration Test and  added Optional in Vehicle Repository

+++ src/main/java/com/nrapendra/garage/controllers/GarageController.java
@@ -4,6 +4,7 @@ import com.nrapendra.garage.models.FreeParkingGarageSpace;
 import com.nrapendra.garage.models.Vehicle;
 import com.nrapendra.garage.models.VehicleMovementStatus;
 import com.nrapendra.garage.models.VehicleParkingLocation;
+import com.nrapendra.garage.repositories.VehicleRepository;
 import com.nrapendra.garage.services.GarageService;
 import com.nrapendra.garage.utils.AppUtil;
 import com.nrapendra.garage.utils.UrlUtil;
@@ -16,6 +17,8 @@ import org.springframework.web.bind.annotation.*;
 import java.io.IOException;
 import java.text.ParseException;
 
+import static com.nrapendra.garage.utils.AppUtil.VEHICLE_DOES_NOT_EXIST;
+
 /**
  * Created by NrapendraKumar
  */
@@ -27,27 +30,37 @@ public class GarageController extends OpenAPIController {
 
     private final GarageService garageService;
 
+    private final VehicleRepository vehicleRepository;
+
     @RequestMapping(value = UrlUtil.GARAGE_ENTER_VEHICLE_URL, method = RequestMethod.POST)
     public ResponseEntity<VehicleMovementStatus> enterVehicle(@RequestBody Vehicle vehicle) throws IOException, ParseException {
         VehicleMovementStatus vehicleMovementStatus = garageService.addVehicle(vehicle);
-        return new ResponseEntity<VehicleMovementStatus>(vehicleMovementStatus, HttpStatus.OK);
+        return new ResponseEntity<>(vehicleMovementStatus, HttpStatus.OK);
     }
 
     @RequestMapping(value = UrlUtil.GARAGE_EXIT_VEHICLE_URL, method = RequestMethod.DELETE)
     public ResponseEntity<VehicleMovementStatus> exitVehicle(@PathVariable(AppUtil.VEHICLE_NUMBER) String vehicleNumber) throws IOException, ParseException {
+
+        var vehicle = vehicleRepository.findByVehicleNumber(vehicleNumber);
+
+        if(vehicle.isEmpty()){
+            var vms=new VehicleMovementStatus();
+            vms.setStatus(VEHICLE_DOES_NOT_EXIST);
+            return new ResponseEntity<>(vms, HttpStatus.NOT_FOUND);
+        }
         VehicleMovementStatus vehicleMovementStatus = garageService.removeVehicle(vehicleNumber);
-        return new ResponseEntity<VehicleMovementStatus>(vehicleMovementStatus, HttpStatus.OK);
+        return new ResponseEntity<>(vehicleMovementStatus, HttpStatus.OK);
     }
 
     @RequestMapping(value = UrlUtil.GARAGE_GET_VEHICLE_LOCATION_URL, method = RequestMethod.GET)
     public ResponseEntity<VehicleParkingLocation> getVehicleLocation(@PathVariable(AppUtil.VEHICLE_NUMBER) String vehicleNumber) throws IOException, ParseException {
         VehicleParkingLocation vehicleParkingLocation = garageService.getVehicleParkingLocation(vehicleNumber);
-        return new ResponseEntity<VehicleParkingLocation>(vehicleParkingLocation, HttpStatus.OK);
+        return new ResponseEntity<>(vehicleParkingLocation, HttpStatus.OK);
     }
 
     @RequestMapping(value = UrlUtil.GARAGE_FREE_SPACE_URL, method = RequestMethod.GET)
     public ResponseEntity<FreeParkingGarageSpace> getAllFreeSpaces() throws IOException, ParseException {
         FreeParkingGarageSpace freeParkingGarageSpace = garageService.getFreeParkingLotsInformation();
-        return new ResponseEntity<FreeParkingGarageSpace>(freeParkingGarageSpace, HttpStatus.OK);
+        return new ResponseEntity<>(freeParkingGarageSpace, HttpStatus.OK);
     }
 }
